requirements.txt
Flask
google-cloud-speech
google-genai
flask-cors